// src/react/redux/types/ui/dashboardVisualization.ts
export const DataVisualizationsList = [TableKey, TimeSeriesKey, HistogramKey, Histogram3DKey] as const;
export type DataVisTypes = (typeof DataVisualizationsList)[number];
// ...
type VisIndexMap = {
  [TableKey]: null;
  [TimeSeriesKey]: {
    index: {
      name: string;
      units: string;
      type: "Double" | "Integer";
      dataBase64: string;
    };
  };
  [HistogramKey]: {
    index: {
      name: string;
      units: string;
      type: "Double" | "Integer";
      dataBase64: string;
    };
  };
  [Histogram3DKey]: {
    index: {
      name: string;
      units: string;
      type: "Double" | "Integer";
      dataBase64: string;
    };
  };
};

type VisDataSetDataMap = {
  [TableKey]: {
    $format: "csv";
    data: string;
  };
  [TimeSeriesKey]: {
    index: {
      name: string;
      units: string;
      type: "Double" | "Integer";
      dataBase64: string;
    };
    columns: {
      name: string;
      units: string;
      type: "Double" | "Integer";
      dataBase64: string;
    }[];
  };
  [HistogramKey]: {
    index: {
      name: string;
      units: string;
      type: "Double" | "Integer";
      dataBase64: string;
    };
    histograms: {
      chanName: string;
      type: "h2axis";
      data: {
        x: {
          start: string;
          midPoint: string;
          end: string;
          label: string;
        }[];
        xUnits: string;
        xLabel: string;
        y: {
          value: string;
        }[];
        yUnits: string;
        yLabel: string;
      };
    }[];
  };
  [Histogram3DKey]: {
    index: {
      name: string;
      units: string;
      type: "Double" | "Integer";
      dataBase64: string;
    };
    histograms: {
      chanName: string;
      type: "h3axis";
      data: {
        x: {
          start: string;
          midPoint: string;
          end: string;
          label: string;
        }[];
        xUnits: string;
        xLabel: string;
        y: {
          start: string;
          midPoint: string;
          end: string;
          label: string;
        }[];
        yUnits: string;
        yLabel: string;
        z: {
          value: string;
        }[];
        zUnits: string;
        zLabel: string;
      };
    }[];
  };
};
// ...
export const VisZodSchemaMap: VisZodSchemaMapType = {
  [TableKey]: {
    index: z.null(),
    dataSetSchema: getDataSetSchema(TableKey),
    dataSetDataSchema: z.object({
      $format: z.enum(["csv"]),
      data: z.string()
    })
  },
  [TimeSeriesKey]: {
    index: z.object({
      index: z.object({
        name: z.string(),
        units: z.string(),
        type: z.enum(["Double", "Integer"]),
        dataBase64: z.string()
      })
    }),
    dataSetSchema: getDataSetSchema(TimeSeriesKey),
    dataSetDataSchema: z.object({
      index: z.object({
        name: z.string(),
        units: z.string(),
        type: z.enum(["Double", "Integer"]),
        dataBase64: z.string()
      }),
      columns: z.array(
        z.object({
          name: z.string(),
          units: z.string(),
          type: z.enum(["Double", "Integer"]),
          dataBase64: z.string()
        })
      )
    })
  },
  [HistogramKey]: {
    index: z.object({
      index: z.object({
        name: z.string(),
        units: z.string(),
        type: z.enum(["Double", "Integer"]),
        dataBase64: z.string()
      })
    }),
    dataSetSchema: getDataSetSchema(HistogramKey),
    dataSetDataSchema: z.object({
      index: z.object({
        name: z.string(),
        units: z.string(),
        type: z.enum(["Double", "Integer"]),
        dataBase64: z.string()
      }),
      histograms: z.array(
        z.object({
          chanName: z.string(),
          type: z.literal("h2axis"),
          data: z.object({
            x: z.array(
              z.object({
                start: z.string(),
                midPoint: z.string(),
                end: z.string(),
                label: z.string()
              })
            ),
            xUnits: z.string(),
            xLabel: z.string(),
            y: z.array(
              z.object({
                value: z.string()
              })
            ),
            yUnits: z.string(),
            yLabel: z.string()
          })
        })
      )
    })
  },
  [Histogram3DKey]: {
    index: z.object({
      index: z.object({
        name: z.string(),
        units: z.string(),
        type: z.enum(["Double", "Integer"]),
        dataBase64: z.string()
      })
    }),
    dataSetSchema: getDataSetSchema(Histogram3DKey),
    dataSetDataSchema: z.object({
      index: z.object({
        name: z.string(),
        units: z.string(),
        type: z.enum(["Double", "Integer"]),
        dataBase64: z.string()
      }),
      histograms: z.array(
        z.object({
          chanName: z.string(),
          type: z.literal("h3axis"),
          data: z.object({
            x: z.array(
              z.object({
                start: z.string(),
                midPoint: z.string(),
                end: z.string(),
                label: z.string()
              })
            ),
            xUnits: z.string(),
            xLabel: z.string(),
            y: z.array(
              z.object({
                start: z.string(),
                midPoint: z.string(),
                end: z.string(),
                label: z.string()
              })
            ),
            yUnits: z.string(),
            yLabel: z.string(),
            z: z.array(
              z.object({
                value: z.string()
              })
            ),
            zUnits: z.string(),
            zLabel: z.string()
          })
        })
      )
    })
  }
};
// ...
type VisDataSetZodType<T> = T extends DataVisTypes ? ZodType<VisDataSetType<T>, ZodTypeDef, VisDataSetType<T>> : never;
// ...
export type VisDataSetDataType<T extends DataVisTypes> = VisDataSetDataMap[T];
export type VisIndexMapType<T extends DataVisTypes> = VisIndexMap[T];

type VisZodSchemaMapType = {
  [K in DataVisTypes]: {
    dataSetSchema: VisDataSetZodType<K>;
    dataSetDataSchema: ZodType<VisDataSetDataType<K>>;
    index: ZodType<VisIndexMapType<K>>;
  };
};